+++ mathics/builtin/string/operations.py
@@ -286,7 +286,7 @@ class StringInsert(Builtin):
 
         # Check and create list of position
         listpos = []
-        if pos.has_form(SymbolList, None):
+        if pos.get_head() is SymbolList:
             leaves = pos.get_leaves()
             if not leaves:
                 return strsource
@@ -303,7 +303,7 @@ class StringInsert(Builtin):
             listpos.append(py_pos)
 
         # Check and perform the insertion
-        if strsource.has_form(SymbolList, None):
+        if strsource.get_head() is SymbolList:
             py_strsource = [sub.get_string_value() for sub in strsource.leaves]
             if any(sub is None for sub in py_strsource):
                 return evaluation.message("StringInsert", "strse", Integer1, exp)
@@ -501,7 +501,7 @@ class StringPosition(Builtin):
             overlap = False  # unknown options are teated as False
 
         # convert patterns
-        if patt.has_form(SymbolList, None):
+        if patt.get_head() is SymbolList:
             patts = patt.get_leaves()
         else:
             patts = [patt]
@@ -514,7 +514,7 @@ class StringPosition(Builtin):
         compiled_patts = [re.compile(re_patt) for re_patt in re_patts]
 
         # string or list of strings
-        if string.has_form(SymbolList, None):
+        if string.get_head() is SymbolList:
             py_strings = [s.get_string_value() for s in string.leaves]
             if None in py_strings:
                 return
@@ -770,7 +770,7 @@ class StringRiffle(Builtin):
         if len(separators) > 1:
             return evaluation.message("StringRiffle", "mulsep")
         elif len(separators) == 1:
-            if separators[0].has_form(SymbolList, None):
+            if separators[0].get_head() is SymbolList:
                 if len(separators[0].leaves) != 3 or any(
                     not isinstance(s, String) for s in separators[0].leaves
                 ):
@@ -779,10 +779,10 @@ class StringRiffle(Builtin):
                 return evaluation.message("StringRiffle", "string", Integer(2), exp)
 
         # Validate list of string
-        if not liststr.has_form(SymbolList, None):
+        if not liststr.get_head() is SymbolList:
             evaluation.message("StringRiffle", "list", Integer1, exp)
             return evaluation.message("StringRiffle", "argmu", exp)
-        elif any(leaf.has_form(SymbolList, None) for leaf in liststr.leaves):
+        elif any(leaf.get_head() is SymbolList for leaf in liststr.leaves):
             return evaluation.message("StringRiffle", "sublist")
 
         # Determine the separation token
@@ -790,7 +790,7 @@ class StringRiffle(Builtin):
         if len(separators) == 0:
             sep = " "
         else:
-            if separators[0].has_form(SymbolList, None):
+            if separators[0].get_head() is SymbolList:
                 left = separators[0].leaves[0].value
                 sep = separators[0].leaves[1].value
                 right = separators[0].leaves[2].value
@@ -897,7 +897,7 @@ class StringSplit(Builtin):
                 "StringSplit", "strse", Integer1, Expression("StringSplit", string)
             )
 
-        if patt.has_form(SymbolList, None):
+        if patt.get_head() is SymbolList:
             patts = patt.get_leaves()
         else:
             patts = [patt]