+++ CHANGES.rst
@@ -14,7 +14,9 @@ Internals
 * Module mathics.builtin.box is included as a builtin
 * Extra sort() was removed from Plus.apply
 * Reorganize Definitions.get_definition to deal with heavier rare cases later.
-* Avoid call Definitions.get_definition in favor of Definitions.definition_cache.get always it is possible.  
+* Avoid call Definitions.get_definition in favor of Definitions.definition_cache.get always it is possible.
+* For ``Atoms`` ``get_head`` and ``get_head_name`` were speed-up by using a class member ``_head_symbol`` instead of looking for a symbol each time.
+  
 
 4.0.1
 -----